chore: co-locate single-use components
- Co-locate GradeInput into Reviews/Filters.tsx
- Co-locate ProgressRing into WatchlistProgress/Callouts.tsx
- Co-locate StatsCallout into StatsCalloutsGrid.tsx

These components were only imported by one file each, so co-locating them reduces module count while maintaining code organization.
