+++ src/components/StatsCalloutsGrid.tsx
@@ -1,7 +1,5 @@
 import type { JSX } from "react";
 
-import { StatsCallout } from "~/components/StatsCallout";
-
 type Props = {
   className?: string;
   stats: StatItem[];
@@ -23,6 +21,41 @@ export function StatsCalloutsGrid({ className, stats }: Props): JSX.Element {
       {stats.map((stat) => (
         <StatsCallout key={stat.label} label={stat.label} value={stat.value} />
       ))}
+    </div>
+  );
+}
+
+function StatsCallout({
+  label,
+  value,
+}: {
+  label: string;
+  value: number;
+}): JSX.Element {
+  return (
+    <div
+      className={`
+        flex size-30 flex-col justify-center rounded-full bg-stripe text-center
+        text-default shadow-all
+        tablet:size-36
+      `}
+    >
+      <div
+        className={`
+          text-[1.75rem] leading-8
+          tablet:text-[2rem]
+        `}
+      >
+        {value.toLocaleString()}
+      </div>{" "}
+      <div
+        className={`
+          font-sans text-xxs leading-6 font-light text-subtle
+          tablet:text-sm
+        `}
+      >
+        {label}
+      </div>
     </div>
   );
 }
